Create makejson.html
This is the tool I used to generate the new json values for the different days and the combined file.

Unfortunately because I chose to do this with javascript (because it was easy) it means that there is no way to automatically save the new json values to file. The way I saved the new values was to copy them out of the console in chrome and paste them into the new file.

(newMEGAData needs to be written to console manually after the page has loaded using the same syntax that the other variables are echoed to console)

The files that it loads need to be saved from the octranspo site using a browser web tool to save each day's json response as you load them. OCTranspo seems to serve them from a uniquely generate url so I can't think of a way to do this automatically.